mtd: onenand: samsung: Unify resource order for controller variants
Before this patch, the order of memory resources requested by the driver
was controller base as first and OneNAND chip base as second for
S3C64xx/S5PC100 variant and the opposite for S5PC110/S5PV210 variant.

To make this more consistent, this patch swaps the order of resources
for the latter and updates platform code accordingly. As a nice side
effect there is a slight reduction in line count of probe function.

Signed-off-by: Tomasz Figa <tomasz.figa@gmail.com>
Signed-off-by: Paweł Chmiel <pawel.mikolaj.chmiel@gmail.com>

---

In the first review, it was suggested to split this patch in two due
to the renaming of fields in the main onenand structure.  However,
it is only fields in the private s3c_onenand structure that are being
renamed so it has been left as one patch.