drm: g2d: Add support for older version 3.0 variant
The 3.0 variant of G2D is present on S5PV210 and Exynos 4210.  It
differs from v4.1 in several ways
- No DMA submission of commands
- No blend mode option
- No solid fill regs
- No YCrCb support
- Only some of the AXI configuration
- Different scaling methods
- Extra cache controls

Of the above, only the DMA submission of commands is actually implemented
kernel side, the rest is in userspace.  Note that there should be verification
of input registers and that they exist, but this is currently not done.

Instead of the DMA submission of command, we keep track of where we
are in the cmdlists and write the values directly to the registers.

It has been tested on an Samsung SGH-T959P with a modified libdrm

Signed-off-by: Jonathan Bakker <xc-racer2@live.ca>