dts: s5pv210: Set clock as syscon compatible as well
The pinctrl driver needs to access the registers of the clock for the
setting of the external interrupts.  It previously was simply mapping
these registers, but then was trying to use the regmap interface on
them.  To avoid having to handle s5pv210 eint setting separately, simply
allow the clock registers to be mapped as a syscon regmap.

Signed-off-by: Jonathan Bakker <xc-racer2@live.ca>