sound: soc: samsung: Add aries_wm8994 machine driver
Aries boards have a relatively complex audio setup with a WM8994
connected to I2S0.  Connected to the WM8994 is also the BT chip
as well as a modem.

Jack detection is done by a combination of GPIOs and an ADCfor the
headphone jack) along with an extcon device for the USB audio.

Signed-off-by: Jonathan Bakker <xc-racer2@live.ca>